requirements.txt
flask
requests
newrelic
Flask-PyMongo
Flask-Mail
celery
redis

## Production ##
gunicorn
